Alllow PHP 8.4 (#62)
PHP 8.4 support.

This PR adds official support for PHP 8.4.

There was a change necessary in the package, which requires loading WordPress 6.7 as stubs instead of `latest` (6.8 - May 2025) to not cause static analyzers to fail in PHP 7.4 due to a syntax error. 
This will be reverted in the future when https://github.com/php-stubs/wordpress-stubs/issues/277 is fixed and released.